requirements.txt
asgiref==3.4.1
astroid==2.7.3
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-extensions==3.1.3
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
httpie==2.5.0
idna==3.2
importlib-metadata==4.8.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.6
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.10.2
pyparsing==2.4.7
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
zipp==3.6.0
